requirements.txt
flask
requests
flask_sqlalchemy
flask_migrate
flask_script
flask_cors
pymysql
# redis
jsonschema
pillow
bcrypt
PyJWT
pycryptodome